Re-add replay launching logic
Removed in previous commit as part of removing selection handlers. This is common lobby logic that is important to keep in the Core

Also implemented access control for functions/properties of SpringProcessController.

Launch games are now throwing functions, throwing an error if Spring is already running when the user asks to start a new game. (This is not an inherent limitation of the engine, and thus it should be considered whether this restriction should remain.)