Data Restructure for Correctness
Further restructuring is required to correct remaining issues, such as:
- Inconsistent responsibilities between `Client`, `Connection`, and `Socket`
- `AccountInfoController` standing as a member of `Client`, not `AuthenticatedClient`
- Unclear dependency relationships between `ReplayController`, `DownloadsController`, and `ResourceManager`.

This commit corrects a number of structural flaws. Primarily:
1. Information about user’s authentication not being a strict subset of the data describing the connection with a server

Helper methods have been added to the implementation of `Client` to minimise code repetition when dealing with this new structure. `TASServer` has been renamed to `Connection` (Possibly temporarily) to provide greater clarity of its role.

2. The state provided by the server as described after login not being a strict subset of the connection with the server & its authenticated session.

State management has been extracted from `Client` and delegated to a new data structure, `AuthenticatedClient`. This is intended to be contained as a subset of the `Connection` object.

3. `Client` state updates dictating implementation details of the user interface (or lack of).

`WindowManager` & `ClientWindowManager` have been removed, and replaced with implementations of `UpdateNotifier` that chain updates to any object that must perform some action or update in response.