Refactor in preparation for Tachyon support
Dedicates a class to each lobby protocol, allowing it to decode the messages from data on its own. Instead of using the ServerProtocol enum, we hand in a Lobby object that knows how to detect data (or in the case of Tachyon servers, we'll have to give the correct lobby object directly to it).

It's up to implementations to handle the decision of which Lobby object to give us.