Better Unitsync Error Handling (`Optional.throwIfNil`)
Allows throwing errors for group handling where individual handling isn't useful, complete with recording the line and file where it failed. 

Also changes engine search directory to the engines folder: this potentially breaks, for example, searching in the Applications folder on MacOS, so some nice alternative to this needs to be established.